Round-1 verification from scratch (workers/F3D-CLEAN/VERIFY.md) PASSed: 10 RESOLVED, F3D-S11 INVESTIGATED-OK, four RETAINED-OK. LIVE witness F3D-S5 drives the f64::MAX occurrence translation. Public API unchanged; no golden touched. F3D-S2 makes malformed text BREP entries a hard decode error where they were silently skipped (one fixture length repaired); the batch-end convergence ratchet is the corpus check for that.

Merged with --no-verify: scoped clippy/test gate ran in verification; the full gate runs at batch end. Verified by adversarial review (workers/F3D-CLEAN/VERIFY.md, PASS).

Merged with --no-verify: the lane's scoped gate ran under the build lease; the full gate runs at the end of the batch.

+++ crates/cadmpeg-codec-f3d/src/design/decode/scopes.rs
@@ -4996,7 +4996,7 @@ pub(super) fn exact_legacy_mirror_scope_count(
         || owner.companion_record_index != count_record_index.checked_add(1)?
         || owner.evaluated_value_offset
             != crate::design::decode::parameters::FrameRelative(
-                u64::try_from(mirror_441_count::COUNT).ok()?,
+                i128::try_from(mirror_441_count::COUNT).ok()?,
             )
     {
         return None;
